requirements.txt
pyautogen
openai
azure-identity
azure-kusto-data
pyyaml